requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
pytest
httpx
python-jose
passlib
python-jose[cryptography]
passlib[bcrypt]